Improve KeyboardInterrupt handling.
Solve an issue where select() gets a local interface and a remote connection ready, tries to send to the remote connection, notices that it's down and NULLs it out, then it doesn't detect that it's dealing with that remote connection later on. Solve by reversing the ordering of 'additionalListeners' in the select inputReady array.